Complete adding dollars_saved and carbon_saved to activities creation

+++ db/migrate/20230104142905_create_activities.rb
@@ -7,7 +7,7 @@ class CreateActivities < ActiveRecord::Migration[5.2]
       t.integer :num_drinks
       t.string :drink_type
       t.float :dollars_saved
-      t.integer :lbs_carbon_saved
+      t.float :lbs_carbon_saved
       t.references :user, foreign_key: true
 
       t.timestamps
